Move logic for selected path restoration to data_model file
